test(gateway): extract session roster cache fixtures
