+++ src/gateway/server-methods/sessions-read-cache.test-support.ts
@@ -1,0 +1,58 @@
+import { expect, vi } from "vitest";
+import type { SessionsListParams } from "../../../packages/gateway-protocol/src/index.js";
+import type { OpenClawConfig } from "../../config/types.openclaw.js";
+import type { GatewaySessionRow } from "../session-utils.types.js";
+import { sessionReadHandlers } from "./sessions-read.js";
+import type { GatewayClient, GatewayRequestContext, RespondFn } from "./types.js";
+
+export { sessionReadHandlers };
+
+export function identifiedClient(profileId: string): GatewayClient {
+  return {
+    connect: {
+      minProtocol: 1,
+      maxProtocol: 1,
+      client: { id: "openclaw-control-ui", version: "test", platform: "test", mode: "webchat" },
+      role: "operator",
+      scopes: ["operator.read", "operator.write"],
+    },
+    authenticatedUserProfile: {
+      profileId,
+      displayName: profileId,
+      hasAvatar: false,
+      updatedAt: 1,
+    },
+  };
+}
+
+export function requestContext(config: OpenClawConfig): GatewayRequestContext {
+  return {
+    chatAbortControllers: new Map(),
+    getRuntimeConfig: () => config,
+    getSessionEventSubscriberConnIds: () => new Set(),
+    loadGatewayModelCatalog: async () => [],
+    logGateway: { debug: vi.fn() },
+  } as unknown as GatewayRequestContext;
+}
+
+export async function listSessions(params: {
+  client: GatewayClient;
+  context: GatewayRequestContext;
+  request: SessionsListParams;
+}) {
+  const responses: Parameters<RespondFn>[] = [];
+  await sessionReadHandlers["sessions.list"]?.({
+    params: params.request,
+    client: params.client,
+    context: params.context,
+    respond: (...response: Parameters<RespondFn>) => responses.push(response),
+  } as never);
+  expect(responses).toHaveLength(1);
+  expect(responses[0]?.[0]).toBe(true);
+  return responses[0]?.[1] as {
+    count: number;
+    nextOffset: number | null;
+    sessions: GatewaySessionRow[];
+    totalCount: number;
+  };
+}
